Hide snooze timer for edge cases; unsnooze for no alarms

+++ Alarms.h
@@ -141,7 +141,7 @@ public:
   void update(const AlarmLevel& alarm_level);
 
   // Get snooze time remaining
-  int getRemainingSnoozeTime();
+  unsigned long getRemainingSnoozeTime();
 
 private:
   const int beeper_pin_;
